% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{subsetRDS}
\alias{subsetRDS}
\title{subsetRDS}
\usage{
subsetRDS(rds, group = "all", sample = "all", cluster = "all", verbose = TRUE)
}
\arguments{
\item{rds}{seurat object}

\item{group}{subset group,default:all}

\item{sample}{subset sample,default:all}

\item{cluster}{subset cluster,default:all}
}
\value{
Return a seurat object.
}
\description{
subsetRDS
}
