% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{subsetSeurat}
\alias{subsetSeurat}
\title{subsetSeurat}
\usage{
subsetSeurat(
  rds,
  meta,
  value = "all",
  slot = "data",
  compair = ">",
  verbose = TRUE
)
}
\arguments{
\item{rds}{seurat object}

\item{meta}{choose meta sub(cluster,sample,group,genename .etc)}

\item{value}{You can select the type to be filtered in the meta.When a gene is selected by meta, fill in the threshold of gene expression here}

\item{slot}{when meta choose a gene,you can use slot parameter,it can choose "counts","data"}

\item{compair}{when meta choose a gene,you can use compair parameter,it can choose ">","<","=",">=","<="}
}
\value{
Return a seurat object.
}
\description{
subsetSeurat
}
